TAT-183: Join create/edit feature flows.
THe edit flow presupposes the create flow was executed just before it.
If they're kept separate, the edit flow might be run standalone and fail.

The more I work on these, the greater the coupling between them, which buys us execution speed and easier maintenance.

The manual test flow for these scenarios is also tightly coupled, so I decided to join them in code as well.