Refactor TAT-389: Remove thin-wrapping locators around e.g. anotherLocator().get(i).
While I've written many such locators before, current practice is to avoid them if possible, because from experience I've found them to be rarely needed.

This kind of encapsulation is still found elsewhere, but only where it's proven necessary, or where it's not yet proven to be so rarely needed like this one.