TAT-389, fix: Compare arrays instead of checking if values are contained.
The relevance of this is that the test uses the same data for multiple entries (e.g. same submitted date 4 times).

By iterating over an array of 4 equal expected values and expecting each of them to be contained in the array with data loaded from the UI, the test doesn't really make sure all the data is there.
The second, third, and fourth dates may be completely wrong, but tests would pass anyway.

Actually comparing the arrays fixes that in this case.